Transformers in TransformerTask and MappingTransformer use the same declaration
